+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.13.2] 2025-04-30
+## [1.13.3] 2025-05-03
+### Changed
+- The input directory in the gui is now a file input instead of a text input.
+
+### Removed
+- Removed the `readme.md` file in the sesh folder.
+
+## [1.13.2] 2025-05-03
 ### Added
 - There is now an "Items per page option" in the settings.
 - Added an on hover effect to the Activity Heatmap